scripts/requirements.txt
gitpython
openstacksdk
pytest